% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{Encode DNA sequences into position-wise nucleotide factors}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A character vector containing one or more DNA sequences.}
}
\value{
A data frame where each column represents a nucleotide position.
}
\description{
This function converts DNA strings (e.g. "ATCGA") into a data frame,
where each column corresponds to a nucleotide position, and each value
is a factor with levels A, T, C, G. This encoding is useful for machine
learning model inputs.
}
\examples{
dna_encoding(c("ATCGA", "GGTAC"))
}
