% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Predict m6A methylation probability and status for multiple samples}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{A trained random forest model object.}

\item{feature_df}{A data frame containing features such as gc_content,
RNA_type, RNA_region, exon_length, distance_to_junction,
evolutionary_conservation, and DNA_5mer.}

\item{positive_threshold}{A numeric threshold (0–1) that determines whether
a site is classified as "Positive" (above threshold) or "Negative". Default value is 0.5.}
}
\value{
A data frame containing the original features and two new columns:
predicted_m6A_prob and predicted_m6A_status.
}
\description{
This function takes a trained random forest model and a feature data frame
as input, and outputs predicted m6A probabilities and classification results
("Positive" or "Negative") for each sample.
}
\examples{
# Load data objects included in the package
rf_fit <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
example_data <- read.csv(system.file("extdata", "m6A_input_example.csv", package = "m6APrediction"))

# Run prediction example
prediction_multiple(rf_fit, example_data, 0.6)
}
