% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Predict m6A methylation probability and status for a single RNA site}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold
)
}
\arguments{
\item{ml_fit}{A trained random forest model object.}

\item{gc_content}{Numeric GC content value between 0 and 1.}

\item{RNA_type}{RNA type as a string ("mRNA", "lincRNA", "lncRNA", "pseudogene").}

\item{RNA_region}{RNA region as a string ("CDS", "intron", "3'UTR", "5'UTR").}

\item{exon_length}{Numeric exon length value.}

\item{distance_to_junction}{Numeric distance from exon junction.}

\item{evolutionary_conservation}{Numeric conservation score between 0 and 1.}

\item{DNA_5mer}{A 5-character DNA sequence string (A, T, C, G).}

\item{positive_threshold}{Numeric classification threshold (0–1).Default value is 0.5.}
}
\value{
A named vector with two values:
predicted_m6A_prob and predicted_m6A_status.
}
\description{
This function predicts m6A probability and classification for one
individual observation using user-specified input feature values.
}
\examples{
rf_fit <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
prediction_single(rf_fit, 0.6, "mRNA", "CDS", 12, 5, 0.8, "ATCGAT", 0.5)
}
